remove session provider
