+++ src/app/layout.tsx
@@ -13,7 +13,6 @@ import {
 } from '@mantine/core';
 import { Notifications } from '@mantine/notifications';
 import { Analytics } from '@vercel/analytics/react';
-import { SessionProvider } from 'next-auth/react';
 import { NavigationBar } from '~/components/NavigationBar';
 import { bricolage } from '~/lib/font';
 import theme from '~/lib/mantine';
@@ -30,19 +29,19 @@ function RootLayout({ children }: Props) {
       </head>
 
       <body className={bricolage.className}>
-        <SessionProvider>
-          <MantineProvider theme={theme}>
-            <Notifications />
-            <AppShell header={{ height: 60 }} mih="100%">
-              <AppShellHeader>
-                <NavigationBar />
-              </AppShellHeader>
-              <AppShellMain h="100%">
-                <Box h="100%">{children}</Box>
-              </AppShellMain>
-            </AppShell>
-          </MantineProvider>
-        </SessionProvider>
+        <MantineProvider theme={theme}>
+          <Notifications />
+
+          <AppShell header={{ height: 60 }} mih="100%">
+            <AppShellHeader>
+              <NavigationBar />
+            </AppShellHeader>
+
+            <AppShellMain h="100%">
+              <Box h="100%">{children}</Box>
+            </AppShellMain>
+          </AppShell>
+        </MantineProvider>
 
         <Analytics />
       </body>
