+++ src/components/NavigationBar.tsx
@@ -1,35 +1,22 @@
-'use client';
-
-import {
-  ActionIcon,
-  Anchor,
-  Group,
-  useMantineColorScheme,
-} from '@mantine/core';
-import {
-  IconBox,
-  IconLogout,
-  IconMoonStars,
-  IconSun,
-} from '@tabler/icons-react';
-import { signOut, useSession } from 'next-auth/react';
+import { ActionIcon, Anchor, Group } from '@mantine/core';
+import { IconBox, IconLogout } from '@tabler/icons-react';
+import { signOut } from '~/lib/auth.react';
 import Link from 'next/link';
+import ColorSchemeToggle from './colorSchemeToggle';
+import { auth } from '~/lib/auth';
+import { SignoutButton } from './signoutButton';
 
-export const NavigationBar = () => {
-  const { colorScheme, toggleColorScheme } = useMantineColorScheme();
-  const session = useSession();
-
-  const handleSignOut = async () => {
-    await signOut();
-  };
+export const NavigationBar = async () => {
+  const session = await auth();
 
   return (
     <>
       <Group style={{ height: '100%' }} px={20} justify="space-between">
         <Group>
           <Anchor href="/">
-            <IconBox size={50} />
+            <IconBox size={32} />
           </Anchor>
+
           <Anchor
             component={Link}
             href="/files"
@@ -39,6 +26,7 @@ export const NavigationBar = () => {
           >
             List all files
           </Anchor>
+
           <Anchor
             component={Link}
             href="/files/upload"
@@ -51,22 +39,8 @@ export const NavigationBar = () => {
         </Group>
 
         <Group>
-          {session.status === 'authenticated' && (
-            <ActionIcon variant="default" onClick={handleSignOut}>
-              <IconLogout size={16} />
-            </ActionIcon>
-          )}
-          <ActionIcon
-            variant="default"
-            onClick={() => toggleColorScheme()}
-            size={30}
-          >
-            {colorScheme === 'dark' ? (
-              <IconSun size={16} />
-            ) : (
-              <IconMoonStars size={16} />
-            )}
-          </ActionIcon>
+          {session?.user != null && <SignoutButton />}
+          <ColorSchemeToggle />
         </Group>
       </Group>
     </>
