+++ src/components/colorSchemeToggle.module.css
@@ -1,0 +1,24 @@
+.icon {
+  width: rem(20px);
+  height: rem(20px);
+}
+
+.dark {
+  @mixin dark {
+    display: none;
+  }
+
+  @mixin light {
+    display: block;
+  }
+}
+
+.light {
+  @mixin light {
+    display: none;
+  }
+
+  @mixin dark {
+    display: block;
+  }
+}
